chore: package refactors and left over changes for git autocommit pr (#33768)
## Description
- Adding refactors

## Automation

/ok-to-test tags="@tag.Git"

### :mag: Cypress test results
