+++ app/server/appsmith-server/src/main/java/com/appsmith/server/dtos/AutoCommitTriggerDTO.java
@@ -1,4 +1,4 @@
-package com.appsmith.server.helpers.ce.autocommit;
+package com.appsmith.server.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
