+++ app/server/appsmith-server/src/main/java/com/appsmith/server/git/AutoCommitEventHandler.java
@@ -1,0 +1,3 @@
+package com.appsmith.server.git;
+
+public interface AutoCommitEventHandler extends AutoCommitEventHandlerCE {}
