+++ app/server/appsmith-server/src/main/java/com/appsmith/server/git/AutoCommitEventHandlerCE.java
@@ -1,4 +1,4 @@
-package com.appsmith.server.solutions.ce;
+package com.appsmith.server.git;
 
 import com.appsmith.server.events.AutoCommitEvent;
 import reactor.core.publisher.Mono;
