+++ app/server/appsmith-server/src/main/java/com/appsmith/server/git/AutoCommitEventHandlerImpl.java
@@ -1,4 +1,4 @@
-package com.appsmith.server.solutions;
+package com.appsmith.server.git;
 
 import com.appsmith.external.git.GitExecutor;
 import com.appsmith.server.configurations.ProjectProperties;
@@ -7,7 +7,6 @@ import com.appsmith.server.helpers.DSLMigrationUtils;
 import com.appsmith.server.helpers.GitFileUtils;
 import com.appsmith.server.helpers.RedisUtils;
 import com.appsmith.server.services.AnalyticsService;
-import com.appsmith.server.solutions.ce.AutoCommitEventHandlerCEImpl;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
