+++ app/server/appsmith-server/src/main/java/com/appsmith/server/git/GitRedisUtils.java
@@ -17,17 +17,19 @@ public class GitRedisUtils {
 
     private final RedisUtils redisUtils;
 
-    public Mono<Boolean> addFileLock(String defaultApplicationId) {
+    public Mono<Boolean> addFileLock(String defaultApplicationId, Boolean isRetryAllowed) {
+        long numberOfRetries = Boolean.TRUE.equals(isRetryAllowed) ? MAX_RETRIES : 0L;
+
         return redisUtils
                 .addFileLock(defaultApplicationId)
-                .retryWhen(Retry.fixedDelay(MAX_RETRIES, RETRY_DELAY)
+                .retryWhen(Retry.fixedDelay(numberOfRetries, RETRY_DELAY)
                         .onRetryExhaustedThrow((retryBackoffSpec, retrySignal) -> {
                             throw new AppsmithException(AppsmithError.GIT_FILE_IN_USE);
                         }));
     }
 
-    public Mono<Boolean> addFileLockWithoutRetry(String defaultApplicationId) {
-        return redisUtils.addFileLock(defaultApplicationId);
+    public Mono<Boolean> addFileLock(String defaultApplicationId) {
+        return addFileLock(defaultApplicationId, Boolean.TRUE);
     }
 
     public Mono<Boolean> releaseFileLock(String defaultApplicationId) {
