+++ app/server/appsmith-server/src/main/java/com/appsmith/server/git/autocommit/helpers/AutoCommitEligibilityHelperImpl.java
@@ -1,10 +1,11 @@
-package com.appsmith.server.helpers.ce.autocommit;
+package com.appsmith.server.git.autocommit.helpers;
 
 import com.appsmith.external.annotations.FeatureFlagged;
 import com.appsmith.external.enums.FeatureFlagEnum;
 import com.appsmith.server.constants.ArtifactType;
 import com.appsmith.server.domains.GitArtifactMetadata;
 import com.appsmith.server.domains.Layout;
+import com.appsmith.server.dtos.AutoCommitTriggerDTO;
 import com.appsmith.server.dtos.PageDTO;
 import com.appsmith.server.git.GitRedisUtils;
 import com.appsmith.server.helpers.CommonGitFileUtils;
@@ -54,7 +55,7 @@ public class AutoCommitEligibilityHelperImpl extends AutoCommitEligibilityHelper
                 .defaultIfEmpty(FALSE)
                 .cache();
 
-        return Mono.defer(() -> gitRedisUtils.addFileLockWithoutRetry(defaultApplicationId))
+        return Mono.defer(() -> gitRedisUtils.addFileLock(defaultApplicationId, FALSE))
                 .then(Mono.defer(() -> isServerMigrationRequiredMonoCached))
                 .then(Mono.defer(() -> gitRedisUtils.releaseFileLock(defaultApplicationId)))
                 .then(Mono.defer(() -> isServerMigrationRequiredMonoCached))
